[winpr,collections] ListDictionary
